ref(cloudflare): Prepare for WorkerEntrypoint
